--- test/src/testing/profiling/__init__.py
@@ -1,11 +1,0 @@
-def time_function(f):
-    import utime
-    """From http://docs.micropython.org/en/latest/esp8266/reference/speed_python.html"""
-    myname = str(f).split(' ')[1]
-    def new_func(*args, **kwargs):
-        t = utime.ticks_us()
-        result = f(*args, **kwargs)
-        delta = utime.ticks_diff(utime.ticks_us(), t)
-        print('Function {} Time = {:6.3f}ms'.format(myname, delta/1000))
-        return result
-    return new_func
